feat: Press key 's' to start the game
Starting the game will now open the game window with the text
"Press key 's' to start the game". When the player presses the
s key on the keyboard, the game and the timer will start. This
leads to a similar gaming experience irrespective of the hardware
used.

Refactor external attributes of the TouchGame class into a seperate
ExternalGameAttributes class.